+++ routes/api.php
@@ -1,7 +1,7 @@
 <?php
 
-use App\Http\Controllers\Api\Ewallet;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Api\Ewallet;
 use Illuminate\Support\Facades\Route;
 
 // Route::get('/user', function (Request $request) {
@@ -10,3 +10,13 @@ use Illuminate\Support\Facades\Route;
 
 
 Route::post("register", [Ewallet::class, "register"]);
+
+Route::post("login", [Ewallet::class, "login"]);
+
+Route::group(["middleware" => ["auth:sanctum"]], function () {
+  //profile
+  Route::get("profile", [Ewallet::class, "profile"]);
+
+  //logout
+  Route::get("logout", [Ewallet::class, "logout"]);
+});