+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# E-Wallet Api v1.1.0
+
+-- Created login api, profile api
+-- Created a route for profile and login api
+-- Still developing the logout api
+
 # E-Wallet Api v1.0.0
 
--- Created Register Api
+-- Created Register api